+++ Tests.Common/Rounding/TestIntegerRoundingDown.cs
@@ -13,35 +13,35 @@ namespace Tests.Common.Rounding
 		public void Should_return_Rounded_1000_for_Decimal_with_01_Cents()
 		{
 			decimal testInputs = 1000.01m;
-			int testResult = IntRounding.RoundDown(testInputs);
+			long testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_POS_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_1000_for_Decimal_with_20_Cents()
 		{
 			decimal testInputs = 1000.20m;
-			int testResult = IntRounding.RoundDown(testInputs);
+			long testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_POS_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_1000_for_Decimal_with_50_Cents()
 		{
 			decimal testInputs = 1000.50m;
-			int testResult = IntRounding.RoundDown(testInputs);
+			long testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_POS_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_1000_for_Decimal_with_60_Cents()
 		{
 			decimal testInputs = 1000.60m;
-			int testResult = IntRounding.RoundDown(testInputs);
+			long testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_POS_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_1000_for_Decimal_with_99_Cents()
 		{
 			decimal testInputs = 1000.99m;
-			int testResult = IntRounding.RoundDown(testInputs);
+			long testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_POS_RESULT_NUMBER, testResult);
 		}
 
@@ -51,35 +51,35 @@ namespace Tests.Common.Rounding
 		public void Should_return_Rounded_Negative_1000_for_Decimal_with_01_Cents()
 		{
 			decimal testInputs = -1000.01m;
-			int testResult = IntRounding.RoundDown(testInputs);
+			long testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_NEG_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_Negative_1000_for_Decimal_with_20_Cents()
 		{
 			decimal testInputs = -1000.20m;
-			int testResult = IntRounding.RoundDown(testInputs);
+			long testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_NEG_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_Negative_1000_for_Decimal_with_50_Cents()
 		{
 			decimal testInputs = -1000.50m;
-			int testResult = IntRounding.RoundDown(testInputs);
+			long testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_NEG_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_Negative_1000_for_Decimal_with_60_Cents()
 		{
 			decimal testInputs = -1000.60m;
-			int testResult = IntRounding.RoundDown(testInputs);
+			long testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_NEG_RESULT_NUMBER, testResult);
 		}
 		[Test()]
 		public void Should_return_Rounded_Negative_1000_for_Decimal_with_99_Cents()
 		{
 			decimal testInputs = -1000.99m;
-			int testResult = IntRounding.RoundDown(testInputs);
+			long testResult = IntRounding.RoundDown(testInputs);
 			Assert.AreEqual(TEST_NEG_RESULT_NUMBER, testResult);
 		}
 	}
